agent-service/requirements.txt
google-generativeai>=0.3.0
numpy==1.24.3
python-dotenv>=1.0.0
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
newsapi-python==0.2.7
torch==2.0.1
keybert==0.7.0
httpx>=0.25.1
transformers==4.30.2 